Separate the text library into message and translation libraries.
These segments each do vastly different things, and it is entirely possible to use the translation library without needing the message library. This also makes testing easier as the translation library does not require Sponge to be initialized why the message library does. As such, it makes sense these should be kept separate for organization and reusability. This will also help should the message library be extended to support proper placeholders.