Add plugin library for managing common resources.
The approach for this library is to establish a framework with standardized decision on the location of resources and structure of the plugin. This is necessary in order to automate parts of initialization without requiring a mess of configuration options for each decision (potentially in the future, but not necessary for our projects).